Module Dsa.K_gen

K_gen can be instantiated over a hashing module to obtain an RFC6979 compliant k-generator for that hash.

Parameters

H : Mirage_crypto.Hash.S

Signature

val generate : key:priv -> Cstruct.t -> Cstruct.t

generate ~key digest deterministically takes the given private key and message digest to a k suitable for seeding the signing process.