Module Dsa.Precompute

Operations to precompute useful data meant to be hardcoded in mirage-crypto-ec before compilation

val generator_tables : unit -> string array array array

Return an array of shape (Fe_length * 2, 15, 3) containing multiples of the generator point for the curve. Useful only to bootstrap tables necessary for scalar multiplication.