Module Dsa.Primitive

Low-level arithmetic operations.

val generator : pub

generator is the generator point (base point) of the curve.

val add : pub -> pub -> pub

add p q is the sum of points p and q.

val scalar_mult : priv -> pub -> pub

scalar_mult s p is the scalar multiplication of p by s.